Moved buffer updates and descriptor set updates onto managed threads
Now they run on their own threads, instead of spawning new threads for each update, which chewed up a lot of resources simply spawning new threads.

Also made new thread lock class to help with managing a lock hierarchy to avoid deadlocks. The code currently crashes due to locks in the api manager not being set up correctly, but I wanted to push now because I'm about to change a lot of code and I want to have the repo for reference to the old code.